requirements.txt
# Core dependencies
python-dotenv>=1.0.0
redis>=4.0.0
faster-whisper>=0.10.0
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.21.0
transformers>=4.30.0
huggingface-hub>=0.16.0

# Audio processing
librosa>=0.10.0
sounddevice>=0.4.0
webrtcvad>=2.0.10

# Web framework
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0.0

# Machine Learning
scikit-learn>=1.3.0
pyannote.audio>=3.1.0
diart>=0.9.0

# Utilities
requests>=2.31.0
aiohttp>=3.8.0
